requirements.txt
attrs==21.2.0
click==8.0.3
Flask==2.0.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
packaging==21.0
pluggy==1.0.0
protobuf==3.18.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
toml==0.10.2
Werkzeug==2.0.2
